Refactor: coalition 배열 types로 이동
