+++ constants/agenda/agenda.ts
@@ -28,15 +28,3 @@ export const enum Coalition {
   WINTER = 'WINTER',
   OTHER = 'OTHER',
 }
-
-export const coalitionValues = [
-  Coalition.GUN,
-  Coalition.GON,
-  Coalition.GAM,
-  Coalition.LEE,
-  Coalition.SPRING,
-  Coalition.SUMMER,
-  Coalition.AUTUMN,
-  Coalition.WINTER,
-  Coalition.OTHER,
-];
