+++ types/agenda/coalitionTypes.ts
@@ -1,0 +1,13 @@
+import { Coalition } from 'constants/agenda/agenda';
+
+export const coalitionValues = [
+  Coalition.GUN,
+  Coalition.GON,
+  Coalition.GAM,
+  Coalition.LEE,
+  Coalition.SPRING,
+  Coalition.SUMMER,
+  Coalition.AUTUMN,
+  Coalition.WINTER,
+  Coalition.OTHER,
+];
